requirements.txt
absl-py==0.15.0
anyio==3.5.0
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
asgiref==3.5.0
astimp @ file:///home/PFA/python-module/dist/astimp-0.0.3-cp37-cp37m-linux_x86_64.whl
astor==0.8.1
astunparse==1.6.3
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
Babel==2.9.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
branca==0.4.2
cachetools==5.0.0
certifi==2022.5.18.1
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
click==8.0.4
condacolab==0.1.3
cycler==0.11.0
Cython==0.29.28
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Django==3.2.12
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
fastjsonschema @ file:///tmp/build/80754af9/python-fastjsonschema_1620414857593/work/dist
Flask==2.0.3
flatbuffers==1.12
folium==0.12.1.post1
fonttools==4.29.1
gast==0.2.2
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
image==1.5.33
imageio==2.6.1
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources==5.4.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1645888813068/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.1.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1637175076767/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.0
json5==0.9.6
jsonschema @ file:///tmp/build/80754af9/jsonschema_1650008095244/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1642858610849/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1645024288521/work
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.2.0
mkl-random==1.0.4
mkl-service==2.3.0
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775986548/work
nbclassic==0.3.6
nbclient @ file:///tmp/build/80754af9/nbclient_1650308370114/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1649751911790/work
nbformat @ file:///tmp/build/80754af9/nbformat_1649844918397/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
networkx==2.6.3
nibabel==3.2.2
notebook @ file:///opt/conda/conda-bld/notebook_1651734535442/work
notebook-shim==0.1.0
numpy==1.19.5
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==6.2.1
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1644497866770/work
protobuf==3.20.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzbar==0.1.9
pyzmq==19.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SEG @ file:///home/SEG/python-module/dist/SEG-0.0.1-cp37-cp37m-linux_x86_64.whl
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
sqlparse==0.4.2
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.13.2
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1604105014127/work
tqdm==4.63.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.8
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.12.1
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
